+++ src/bin.rs
@@ -110,11 +110,8 @@ fn main() -> Result<(), Error> {
         return Err(Error("Word list is empty".into()));
     }
 
-    let grid_config = generate_grid_config_from_template_string(
-        word_list,
-        &raw_grid_content,
-        args.min_score as f32,
-    );
+    let grid_config =
+        generate_grid_config_from_template_string(word_list, &raw_grid_content, args.min_score);
 
     let result = find_fill(&grid_config.to_config_ref(), None)
         .map_err(|_| Error("Unfillable grid".into()))?;
